mcp: add diagnostics for bad variable replacements
