+++ src/vs/workbench/contrib/mcp/browser/mcpLanguageFeatures.ts
@@ -3,10 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { computeLevenshteinDistance } from '../../../../base/common/diff/diff.js';
 import { Emitter, Event } from '../../../../base/common/event.js';
 import { markdownCommandLink, MarkdownString } from '../../../../base/common/htmlContent.js';
 import { findNodeAtLocation, Node, parseTree } from '../../../../base/common/json.js';
-import { Disposable, DisposableStore, IDisposable, MutableDisposable } from '../../../../base/common/lifecycle.js';
+import { Disposable, DisposableStore, dispose, IDisposable, MutableDisposable } from '../../../../base/common/lifecycle.js';
 import { IObservable } from '../../../../base/common/observable.js';
 import { isEqual } from '../../../../base/common/resources.js';
 import { Range } from '../../../../editor/common/core/range.js';
@@ -14,22 +15,28 @@ import { CodeLensList, CodeLensProvider, InlayHint, InlayHintList } from '../../
 import { ITextModel } from '../../../../editor/common/model.js';
 import { ILanguageFeaturesService } from '../../../../editor/common/services/languageFeatures.js';
 import { localize } from '../../../../nls.js';
+import { IMarkerData, IMarkerService, MarkerSeverity } from '../../../../platform/markers/common/markers.js';
 import { IWorkbenchContribution } from '../../../common/contributions.js';
+import { IConfigurationResolverService } from '../../../services/configurationResolver/common/configurationResolver.js';
 import { ConfigurationResolverExpression, IResolvedValue } from '../../../services/configurationResolver/common/configurationResolverExpression.js';
-import { IMcpConfigPathsService } from '../common/mcpConfigPathsService.js';
+import { IMcpConfigPath, IMcpConfigPathsService } from '../common/mcpConfigPathsService.js';
 import { mcpConfigurationSection } from '../common/mcpConfiguration.js';
 import { IMcpRegistry } from '../common/mcpRegistryTypes.js';
 import { IMcpService, McpConnectionState } from '../common/mcpTypes.js';
 import { EditStoredInput, RemoveStoredInput, RestartServer, ShowOutput, StartServer, StopServer } from './mcpCommands.js';
 
+const diagnosticOwner = 'vscode.mcp';
+
 export class McpLanguageFeatures extends Disposable implements IWorkbenchContribution {
-	private readonly _cachedMcpSection = this._register(new MutableDisposable<{ model: ITextModel; node: Node } & IDisposable>());
+	private readonly _cachedMcpSection = this._register(new MutableDisposable<{ model: ITextModel; inConfig: IMcpConfigPath; tree: Node } & IDisposable>());
 
 	constructor(
 		@ILanguageFeaturesService languageFeaturesService: ILanguageFeaturesService,
 		@IMcpRegistry private readonly _mcpRegistry: IMcpRegistry,
 		@IMcpConfigPathsService private readonly _mcpConfigPathsService: IMcpConfigPathsService,
 		@IMcpService private readonly _mcpService: IMcpService,
+		@IMarkerService private readonly _markerService: IMarkerService,
+		@IConfigurationResolverService private readonly _configurationResolverService: IConfigurationResolverService,
 	) {
 		super();
 
@@ -49,29 +56,97 @@ export class McpLanguageFeatures extends Disposable implements IWorkbenchContrib
 		this._register(languageFeaturesService.inlayHintsProvider.register(patterns, {
 			onDidChangeInlayHints: _mcpRegistry.onDidChangeInputs,
 			provideInlayHints: (model, range) => this._provideInlayHints(model, range),
-		}
-		));
+		}));
 	}
 
 	/** Simple mechanism to avoid extra json parsing for hints+lenses */
 	private _parseModel(model: ITextModel) {
 		if (this._cachedMcpSection.value?.model === model) {
-			return this._cachedMcpSection.value.node;
+			return this._cachedMcpSection.value;
 		}
 
-		const tree = parseTree(model.getValue());
-		const listener = model.onDidChangeContent(() => this._cachedMcpSection.clear());
-		this._cachedMcpSection.value = { model, node: tree, dispose: () => listener.dispose() };
-		return tree;
-	}
-
-	private _provideCodeLenses(model: ITextModel, onDidChangeCodeLens: () => void): CodeLensList | undefined {
-		const inConfig = this._mcpConfigPathsService.paths.get().find(u => isEqual(u.uri, model.uri));
+		const uri = model.uri;
+		const inConfig = this._mcpConfigPathsService.paths.get().find(u => isEqual(u.uri, uri));
 		if (!inConfig) {
 			return undefined;
 		}
 
-		const tree = this._parseModel(model);
+		const value = model.getValue();
+		const tree = parseTree(value);
+		const listeners = [
+			model.onDidChangeContent(() => this._cachedMcpSection.clear()),
+			model.onWillDispose(() => this._cachedMcpSection.clear()),
+		];
+		this._addDiagnostics(model, value, tree, inConfig);
+
+		return this._cachedMcpSection.value = {
+			model,
+			tree,
+			inConfig,
+			dispose: () => {
+				this._markerService.remove(diagnosticOwner, [uri]);
+				dispose(listeners);
+			}
+		};
+	}
+
+	private _addDiagnostics(tm: ITextModel, value: string, tree: Node, inConfig: IMcpConfigPath) {
+		const serversNode = findNodeAtLocation(tree, inConfig.section ? [...inConfig.section, 'servers'] : ['servers']);
+		if (!serversNode) {
+			return;
+		}
+
+		const getClosestMatchingVariable = (name: string) => {
+			let bestValue = '';
+			let bestDistance = Infinity;
+			for (const variable of this._configurationResolverService.resolvableVariables) {
+				const distance = computeLevenshteinDistance(name, variable);
+				if (distance < bestDistance) {
+					bestDistance = distance;
+					bestValue = variable;
+				}
+			}
+			return bestValue;
+		};
+
+		const diagnostics: IMarkerData[] = [];
+		forEachPropertyWithReplacement(serversNode, node => {
+			const expr = ConfigurationResolverExpression.parse(node.value);
+
+			for (const { id, name, arg } of expr.unresolved()) {
+				if (!this._configurationResolverService.resolvableVariables.has(name)) {
+					const position = value.indexOf(id, node.offset);
+					if (position === -1) { continue; } // unreachable?
+
+					const start = tm.getPositionAt(position);
+					const end = tm.getPositionAt(position + id.length);
+					diagnostics.push({
+						severity: MarkerSeverity.Warning,
+						message: localize('mcp.variableNotFound', 'Variable `{0}` not found, did you mean ${{1}}?', name, getClosestMatchingVariable(name) + (arg ? `:${arg}` : '')),
+						startLineNumber: start.lineNumber,
+						startColumn: start.column,
+						endLineNumber: end.lineNumber,
+						endColumn: end.column,
+						modelVersionId: tm.getVersionId(),
+					});
+				}
+			}
+		});
+
+		if (diagnostics.length) {
+			this._markerService.changeOne(diagnosticOwner, tm.uri, diagnostics);
+		} else {
+			this._markerService.remove(diagnosticOwner, [tm.uri]);
+		}
+	}
+
+	private _provideCodeLenses(model: ITextModel, onDidChangeCodeLens: () => void): CodeLensList | undefined {
+		const parsed = this._parseModel(model);
+		if (!parsed) {
+			return undefined;
+		}
+
+		const { tree, inConfig } = parsed;
 		const serversNode = findNodeAtLocation(tree, inConfig.section ? [...inConfig.section, 'servers'] : ['servers']);
 		if (!serversNode) {
 			return undefined;
@@ -187,12 +262,12 @@ export class McpLanguageFeatures extends Disposable implements IWorkbenchContrib
 	}
 
 	private async _provideInlayHints(model: ITextModel, range: Range): Promise<InlayHintList | undefined> {
-		const inConfig = this._mcpConfigPathsService.paths.get().find(u => isEqual(u.uri, model.uri));
-		if (!inConfig) {
+		const parsed = this._parseModel(model);
+		if (!parsed) {
 			return undefined;
 		}
 
-		const tree = this._parseModel(model);
+		const { tree, inConfig } = parsed;
 		const mcpSection = inConfig.section ? findNodeAtLocation(tree, [...inConfig.section]) : tree;
 		if (!mcpSection) {
 			return undefined;
@@ -214,8 +289,8 @@ export class McpLanguageFeatures extends Disposable implements IWorkbenchContrib
 
 		return { hints, dispose: () => { } };
 
-		function annotateServers(node: Node) {
-			if (node.type === 'string' && typeof node.value === 'string' && node.value.includes(ConfigurationResolverExpression.VARIABLE_LHS)) {
+		function annotateServers(servers: Node) {
+			forEachPropertyWithReplacement(servers, node => {
 				const expr = ConfigurationResolverExpression.parse(node.value);
 				for (const { id } of expr.unresolved()) {
 					const saved = inputs[id];
@@ -223,13 +298,7 @@ export class McpLanguageFeatures extends Disposable implements IWorkbenchContrib
 						pushAnnotation(id, node.offset + node.value.indexOf(id) + id.length, saved);
 					}
 				}
-
-			} else if (node.type === 'property') {
-				// skip the property name
-				node.children?.slice(1).forEach(annotateServers);
-			} else {
-				node.children?.forEach(annotateServers);
-			}
+			});
 		}
 
 		function annotateInputs(node: Node) {
@@ -277,6 +346,19 @@ export class McpLanguageFeatures extends Disposable implements IWorkbenchContrib
 			hints.push(hint);
 			return hint;
 		}
+	}
+}
+
+
+
+function forEachPropertyWithReplacement(node: Node, callback: (node: Node) => void) {
+	if (node.type === 'string' && typeof node.value === 'string' && node.value.includes(ConfigurationResolverExpression.VARIABLE_LHS)) {
+		callback(node);
+	} else if (node.type === 'property') {
+		// skip the property name
+		node.children?.slice(1).forEach(n => forEachPropertyWithReplacement(n, callback));
+	} else {
+		node.children?.forEach(n => forEachPropertyWithReplacement(n, callback));
 	}
 }
